transfer Refresh Tree to utils.ts and add tests
