+++ src/job/jobNodeActions.ts
@@ -21,35 +21,13 @@ import { IZoweUSSTreeNode, IZoweJobTreeNode } from "../api/IZoweTreeNode";
 import { ISession } from "@zowe/imperative";
 
 export async function refreshAllJobs(jobsProvider: IZoweTree<IZoweJobTreeNode>) {
-
+    await Profiles.getInstance().refresh();
     jobsProvider.mSessionNodes.forEach((jobNode) => {
         if (jobNode.contextValue === extension.JOBS_SESSION_CONTEXT) {
             utils.labelHack(jobNode);
             jobNode.children = [];
             jobNode.dirty = true;
-        }
-    });
-
-    await jobsProvider.refresh();
-    await Profiles.getInstance().refresh();
-
-    const allProf = Profiles.getInstance().getProfiles();
-    jobsProvider.mSessionNodes.forEach((jobNode) => {
-    if (jobNode.contextValue === extension.JOBS_SESSION_CONTEXT) {
-        for (const profNode of allProf) {
-            if (jobNode.getProfileName() === profNode.name) {
-                jobNode.getProfile().profile = profNode.profile;
-                const SessionProfile = profNode.profile as ISession;
-                if (jobNode.getSession().ISession !== SessionProfile) {
-                    jobNode.getSession().ISession.user = SessionProfile.user;
-                    jobNode.getSession().ISession.password = SessionProfile.password;
-                    jobNode.getSession().ISession.base64EncodedAuth = SessionProfile.base64EncodedAuth;
-                    jobNode.getSession().ISession.hostname = SessionProfile.hostname;
-                    jobNode.getSession().ISession.port = SessionProfile.port;
-                    jobNode.getSession().ISession.rejectUnauthorized = SessionProfile.rejectUnauthorized;
-                    }
-                }
-            }
+            utils.refreshTree(jobNode);
         }
     });
     await jobsProvider.refresh();
